Tweak: replace querying tokens by vault with direct query erc20S
